ub1404 dev workflow
Tweaks to the recipes to avoid repetition of work, and ub1404 dev,
* let apt cookbook handle apt-update globally
* do not download, configure, make, make install if the package is
already installed
* add guards for file deletion to first check whether file is present
* use docker cookbook for image building and running, to only build if
not alrady built and only run if not already running
* drop mysql table and recreate each time

Also,
* bump Docker cookbook to 4.9.3
* bump mysql cookbook to 8.5.1
* add apt cookbook for better apt-update management
* bump depends versions and add apt
* modify readme with customization instructions
* modify all chef runlists to call apt first in the runlist
* add a vagrantfile for dev of ub1404

+++ chef/cookbooks/metasploitable/recipes/phpmyadmin.rb
@@ -14,6 +14,7 @@ bash "download and extract phpmyadmin" do
     tar xvfz /tmp/phpMyAdmin-3.5.8-all-languages.tar.gz -C /var/www/html
     mv /var/www/html/phpMyAdmin-3.5.8-all-languages /var/www/html/phpmyadmin
   EOH
+  not_if { ::File.exists?('/var/www/html/phpmyadmin') }
 end
 
 cookbook_file 'var/www/html/phpmyadmin/config.inc.php' do
